+++ src/AttributeCollection.php
@@ -1,0 +1,40 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Copyright (c) 2025 Andreas Möller
+ *
+ * For the full copyright and license information, please view
+ * the LICENSE.md file that was distributed with this source code.
+ *
+ * @see https://github.com/ergebnis/attribute-collector
+ */
+
+namespace Ergebnis\AttributeCollector;
+
+final class AttributeCollection
+{
+    /**
+     * @var list<Attribute>
+     */
+    private array $attributes;
+
+    private function __construct(Attribute ...$attributes)
+    {
+        $this->attributes = \array_values($attributes);
+    }
+
+    public static function create(Attribute ...$attributes): self
+    {
+        return new self(...$attributes);
+    }
+
+    /**
+     * @return list<Attribute>
+     */
+    public function toArray(): array
+    {
+        return $this->attributes;
+    }
+}